+++ tests/it/main.rs
@@ -32,7 +32,7 @@ mod cmd_op;
 mod comment_op;
 mod comment_node;
 mod template_parts;
-mod const_string_member;
+mod strings_view;
 mod unresolved;
 mod suppress;
 mod config_repos;